Composition diagram now working again. Added interface promotion.

+++ src/org.hyperflex.compositionmodels.diagram/src/org/hyperflex/compositionmodel/diagram/edit/policies/SystemProvidedInterfItemSemanticEditPolicy.java
@@ -7,13 +7,19 @@ import org.eclipse.gef.commands.Command;
 import org.eclipse.gmf.runtime.diagram.core.commands.DeleteCommand;
 import org.eclipse.gmf.runtime.emf.commands.core.command.CompositeTransactionalCommand;
 import org.eclipse.gmf.runtime.emf.type.core.commands.DestroyElementCommand;
+import org.eclipse.gmf.runtime.emf.type.core.commands.DestroyReferenceCommand;
 import org.eclipse.gmf.runtime.emf.type.core.requests.CreateRelationshipRequest;
 import org.eclipse.gmf.runtime.emf.type.core.requests.DestroyElementRequest;
+import org.eclipse.gmf.runtime.emf.type.core.requests.DestroyReferenceRequest;
+import org.eclipse.gmf.runtime.emf.type.core.requests.ReorientReferenceRelationshipRequest;
 import org.eclipse.gmf.runtime.emf.type.core.requests.ReorientRelationshipRequest;
 import org.eclipse.gmf.runtime.notation.Edge;
 import org.eclipse.gmf.runtime.notation.View;
+import org.hyperflex.compositionmodel.diagram.edit.commands.CompositeSrvProducerPromoteCreateCommand;
+import org.hyperflex.compositionmodel.diagram.edit.commands.CompositeSrvProducerPromoteReorientCommand;
 import org.hyperflex.compositionmodel.diagram.edit.commands.ConnectionCreateCommand;
 import org.hyperflex.compositionmodel.diagram.edit.commands.ConnectionReorientCommand;
+import org.hyperflex.compositionmodel.diagram.edit.parts.CompositeSrvProducerPromoteEditPart;
 import org.hyperflex.compositionmodel.diagram.edit.parts.ConnectionEditPart;
 import org.hyperflex.compositionmodel.diagram.part.CompositionModelVisualIDRegistry;
 import org.hyperflex.compositionmodel.diagram.providers.CompositionModelElementTypes;
@@ -28,7 +34,7 @@ public class SystemProvidedInterfItemSemanticEditPolicy extends
 	 * @generated
 	 */
 	public SystemProvidedInterfItemSemanticEditPolicy() {
-		super(CompositionModelElementTypes.SystemProvidedInterf_3007);
+		super(CompositionModelElementTypes.ROSMsgProducer_3007);
 	}
 
 	/**
@@ -45,6 +51,14 @@ public class SystemProvidedInterfItemSemanticEditPolicy extends
 				DestroyElementRequest r = new DestroyElementRequest(
 						incomingLink.getElement(), false);
 				cmd.add(new DestroyElementCommand(r));
+				cmd.add(new DeleteCommand(getEditingDomain(), incomingLink));
+				continue;
+			}
+			if (CompositionModelVisualIDRegistry.getVisualID(incomingLink) == CompositeSrvProducerPromoteEditPart.VISUAL_ID) {
+				DestroyReferenceRequest r = new DestroyReferenceRequest(
+						incomingLink.getSource().getElement(), null,
+						incomingLink.getTarget().getElement(), false);
+				cmd.add(new DestroyReferenceCommand(r));
 				cmd.add(new DeleteCommand(getEditingDomain(), incomingLink));
 				continue;
 			}
@@ -76,7 +90,11 @@ public class SystemProvidedInterfItemSemanticEditPolicy extends
 	 */
 	protected Command getStartCreateRelationshipCommand(
 			CreateRelationshipRequest req) {
-		if (CompositionModelElementTypes.Connection_4002 == req
+		if (CompositionModelElementTypes.Connection_4001 == req
+				.getElementType()) {
+			return null;
+		}
+		if (CompositionModelElementTypes.SystemProvidedInterfExposed_4003 == req
 				.getElementType()) {
 			return null;
 		}
@@ -88,10 +106,15 @@ public class SystemProvidedInterfItemSemanticEditPolicy extends
 	 */
 	protected Command getCompleteCreateRelationshipCommand(
 			CreateRelationshipRequest req) {
-		if (CompositionModelElementTypes.Connection_4002 == req
+		if (CompositionModelElementTypes.Connection_4001 == req
 				.getElementType()) {
 			return getGEFWrapper(new ConnectionCreateCommand(req,
 					req.getSource(), req.getTarget()));
+		}
+		if (CompositionModelElementTypes.SystemProvidedInterfExposed_4003 == req
+				.getElementType()) {
+			return getGEFWrapper(new CompositeSrvProducerPromoteCreateCommand(
+					req, req.getSource(), req.getTarget()));
 		}
 		return null;
 	}
@@ -109,6 +132,22 @@ public class SystemProvidedInterfItemSemanticEditPolicy extends
 			return getGEFWrapper(new ConnectionReorientCommand(req));
 		}
 		return super.getReorientRelationshipCommand(req);
+	}
+
+	/**
+	 * Returns command to reorient EReference based link. New link target or source
+	 * should be the domain model element associated with this node.
+	 * 
+	 * @generated
+	 */
+	protected Command getReorientReferenceRelationshipCommand(
+			ReorientReferenceRelationshipRequest req) {
+		switch (getVisualID(req)) {
+		case CompositeSrvProducerPromoteEditPart.VISUAL_ID:
+			return getGEFWrapper(new CompositeSrvProducerPromoteReorientCommand(
+					req));
+		}
+		return super.getReorientReferenceRelationshipCommand(req);
 	}
 
 }
